fix tg num

+++ accounts/models.py
@@ -11,7 +11,7 @@ class TelegramAccount(models.Model):
     Draft user model with only Telegram ID. Will be attached to User when the
     user will buy subscription.
     """
-    telegram_id = models.PositiveIntegerField('Telegram ID', unique=True)
+    telegram_id = models.BigIntegerField('Telegram ID', unique=True)
     created_at = models.DateTimeField('Добавлен', auto_now_add=True)
     user = models.OneToOneField(CustomUser, related_name='telegram',
                                 on_delete=models.CASCADE, null=True, blank=True)
